Wire up a basically working remark-abbr plugin
:tada:

A lot of simplifying refactoring could be done, particularly on the
micromark plugin.

It would be nice if we didn't have to pass a custom handler in to
remarkRehype in the tests, but I can't work out how to set it in the
plugin. Maybe there's a PR I could raise there.